Add solr Information Object model
Converted elastic search information object model and all dependent
classes to solr equivalent.

WIP: arSolrPluginUtil still has a couple of Elastica references for
search specific functions that need to be adapted/converted for solr.
QubitSearch has not been converted, so no actual documents are added to
Solr index yet. Changes will need to be made to either search itself or
all of the update/populate methods inside models to add the documents to
the solr index.