Update Bootstrap 5 dependency to current version
Unpin Bootstrap 5 and update it to the current version. Due to breaking
changes introduced in v5.2.0 and v5.3.0, and additional import for maps
was needed, updating links using the heading class to use the
default link colour was needed to be set, and updating the default
background colour variable for cards and list group items needed to be
set.