Fix multi value fields in solr (WIP)
Update solr to only use multivalue fields where they are needed.
TODO:

arSolrPlugin includes a large list of fields that are being used to test
field names for being multivalued. This should be updated to work via
yml.

arSolrMapping also has a change that sets all partial foreign type id's
to be multivalued, which needs to be updated to apply that to only the
required id fields.