Fix DO images in the edit DO page (#1915)
Using a div with appropriate styling classes instead of a section for
the image in the edit DO page header since blank.js filters/removes any
sections that do not have content meant for sections.

Also updating the resource sent to the QubitAcl check to point to the
parents' object instead of the resources' object in the scenario when
that object is empty.