Change creators->eventActors in sample accession CSV
I noticed a bug where duplicate events were being created in the
database when importing accessions using the example accession CSV. This
looks to be related to the code in `csvImportBaseTask::importEvents`.
When the "creators" column is present and a column like "eventTypes" is
also present, then code duplicates the number of events because an event
is added when $version=='2.2' AND when $version=='2.3'.

Changing creators to eventActors makes it so that an event is only added
when $version=='2.3'